refactor(core): rename registry module to tool_registry to mirror skill_registry
