+++ src/core/lib/src/lib.rs
@@ -3,15 +3,15 @@
 //! See `README.md` and `docs/adr/` for design.
 
 mod indexing;
-mod registry;
 mod search;
 mod skill;
 mod skill_indexing;
 mod skill_registry;
 mod tool;
+mod tool_registry;
 mod trace;
 
-pub use registry::{SearchHit, ToolRegistry};
+pub use tool_registry::{SearchHit, ToolRegistry};
 pub use skill::Skill;
 pub use skill_registry::{SkillHit, SkillRegistry};
 pub use tool::Tool;
