+++ src/core/lib/src/tool_registry.rs
